requirements.txt
rasterio
numpy
torch
torchvision
Pillow
tqdm
